+++ projects/ng-dynamic-forms/core/src/lib/model/slider/dynamic-slider.model.spec.ts
@@ -28,7 +28,8 @@ describe("DynamicSliderModel test suite", () => {
         expect(model.type).toEqual(DYNAMIC_FORM_CONTROL_TYPE_SLIDER);
         expect(model.value).toBe(config.value);
         expect(model.vertical).toBe(false);
-        expect(model.disabledUpdates).toBeDefined();
+        expect(model.disabledChanges).toBeDefined();
+        expect(model.valueChanges).toBeDefined();
     });
 
     it("should get additional", () => {